refactor: 関数の整理
