+++ Makefile
@@ -5,8 +5,10 @@ FLAGS	+= -fsanitize=thread -g
 SRCS	= src/main.c \
 		  src/parser/open_file.c \
 		  src/parser/error_msg.c \
-		  src/parser/ft_atof.c \
-		  src/parser/safe_function.c
+		  src/parser/safe_function.c \
+		  src/parser/parser.c \
+		  src/parser/free.c \
+		  src/parser/trgb.c
 		#   src/parser/setter_getter.c
 OBJS	= $(SRCS:.c=.o)
 HEADER	= ./includes
