+++ src/libft/ft_split.c
@@ -6,7 +6,7 @@
 /*   By: tsururukakou <tsururukakou@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 18:54:06 by yotsurud          #+#    #+#             */
-/*   Updated: 2025/03/23 16:56:46 by tsururukako      ###   ########.fr       */
+/*   Updated: 2025/03/23 17:14:12 by tsururukako      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,25 +47,13 @@ static char	*ft_make_string_i(char const *s, char *str)
 	return (tmp);
 }
 
-static void	*free_split(char **result, int i)
-{
-	if (result)
-	{
-		while (result[i])
-			free(result[i--]);	
-	}
-	if (result)
-		free(result);
-	return (NULL);
-}
-
 char	**ft_split(char const *s, char *str)
 {
 	int		i;
 	char	**result;
 	
 	if (!s || !str)
-		print_error_and_exit("split function couldn't receive string");
+		print_error_and_exit("split", "split function couldn't receive string");
 	result = (char **)safe_malloc(ft_count_i(s, str) + 1, sizeof(char *));
 	i = 0;
 	while (*s)
