+++ src/parser/open_file.c
@@ -4,5 +4,5 @@ void open_file(char *file_name, int *fd)
 {
     *fd = open(file_name, O_RDONLY);
     if (*fd == -1)
-        print_error_and_exit("couldn't open file");
+        print_error_and_exit("open_file", "couldn't open file");
 }
